Organise the examples directory into 3 folders, each containing one presentation.
