+++ examples/app/views/interactivePresentation/SubmissionConfirmationView.scala
@@ -1,14 +1,15 @@
-package views
+package views.interactivePresentation
 
-object Question1View {
+object SubmissionConfirmationView {
 
   import dsl.Dsl._
+
   import scalatags.Text.all._
 
   def apply(title: String, description: String, author: String, theme: String) = {
     presentation(title, description, author, theme,
       slide(
-        surveyOf2("Question1", 2, "Are You Paying Attention ?","Yes", "No")
+        title3("Thank you for your answer")
       )
     )
   }
